Revert "chore(core): Refactor runtime and split out tests (#19491)"
This reverts commit ec8e9d4f5bd2c8eed5f086356c1c6dd7c8b40b7f.
