+++ core/encode_decode_test.js
