+++ core/realm.rs
@@ -1,8 +1,9 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
-use super::bindings;
-use crate::error::exception_to_err_result;
+
+use crate::bindings;
 use crate::modules::ModuleCode;
 use crate::ops::OpCtx;
+use crate::runtime::exception_to_err_result;
 use crate::runtime::JsRuntimeState;
 use crate::task::MaskResultAsSend;
 use crate::JsRuntime;
