+++ core/serialize_deserialize_test.js
