+++ core/snapshot_util.rs
